Modifications so that changes made in model_tuning branch work properly

+++ conf-core/src/main/java/org/openepics/discs/conf/ejb/AlignmentEJB.java
@@ -58,7 +58,7 @@ import org.openepics.discs.conf.ui.LoginManager;
         return comps;
     }
 
-    public AlignmentRecord findAlignmentRec(int id) {
+    public AlignmentRecord findAlignmentRec(Long id) {
         return em.find(AlignmentRecord.class, id);
     }
